+++ src/main/java/com/cgi/pscatalog/service/dto/OrderDetHstDTO.java
@@ -3,6 +3,7 @@ package com.cgi.pscatalog.service.dto;
 import java.time.Instant;
 import javax.validation.constraints.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
@@ -22,7 +23,7 @@ public class OrderDetHstDTO implements Serializable {
     private Integer orderQuantity;
 
     @NotNull
-    private Long unitPrice;
+    private BigDecimal unitPrice;
 
     private String createdBy;
 
@@ -66,11 +67,11 @@ public class OrderDetHstDTO implements Serializable {
         this.orderQuantity = orderQuantity;
     }
 
-    public Long getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Long unitPrice) {
+    public void setUnitPrice(BigDecimal unitPrice) {
         this.unitPrice = unitPrice;
     }
